+++ examples/mps3-an536/src/bin/svc.rs
@@ -12,13 +12,18 @@ use semihosting::println;
 ///
 /// It is called by the start-up code in `cortex-m-rt`.
 #[no_mangle]
-pub extern "C" fn kmain() {
-    main();
+pub extern "C" fn boot_core(cpu_id: u32) -> ! {
+    match cpu_id {
+        0 => {
+            main();
+        }
+        _ => panic!("unexpected CPU ID {}", cpu_id),
+    }
 }
 
 /// The main function of our Rust application.
 ///
-/// Called by [`kmain`].
+/// Called by [boot_core].
 pub fn main() -> ! {
     let x = 1;
     let y = x + 1;